+++ ovos_local_backend/backend/__init__.py
@@ -31,12 +31,15 @@ def create_app():
     from ovos_local_backend.backend.precise import get_precise_routes
     from ovos_local_backend.backend.external_apis import get_services_routes
     from ovos_local_backend.backend.admin import get_admin_routes
+    from ovos_local_backend.backend.crud import get_database_crud
+
     app = get_auth_routes(app)
     app = get_device_routes(app)
     app = get_stt_routes(app)
     app = get_precise_routes(app)
     app = get_services_routes(app)
     app = get_admin_routes(app)
+    app = get_database_crud(app)
 
     @app.route("/", methods=['GET'])
     @noindex
